start moving stuff around
