+++ Dockerfile
@@ -16,9 +16,9 @@ FROM public.ecr.aws/lambda/python:3.9 as python-run-stage
 COPY --from=python-build-stage /usr/local/lib/python3.9/site-packages /usr/local/lib/python3.9/site-packages
 ENV PYTHONPATH=/usr/local/lib/python3.9/site-packages
 
-COPY app/handler.py ${LAMBDA_TASK_ROOT}
-COPY app/groups_read.py ${LAMBDA_TASK_ROOT}
-COPY app/groups_write.py ${LAMBDA_TASK_ROOT}
+COPY handler.py ${LAMBDA_TASK_ROOT}
+COPY groups_read.py ${LAMBDA_TASK_ROOT}
+COPY groups_write.py ${LAMBDA_TASK_ROOT}
 
 
 CMD [ "handler.handler" ]
